requirements.txt
cassandra-driver==3.20.2
prettytable
requests
pyopenssl
pandas
